python: fix inconsistent naming of Trade enum sets.
They should be named after the plural EnumSet, not the C++ enum. That
was already done for the enums in the primitives library as well as all
shader flags, but not here. They should all also contain a NONE value
for an empty set.

Breaking change, sorry. To avoid similar mistakes in the future, this is
now documented in the API Conventions page.